backend/requirements.txt
asgiref==3.8.1
Django==4.2.9
django-allauth==65.0.2
django-cors-headers==4.3.1
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
